requirements.txt
selenium==3.8.1
Appium-Python-Client==0.25
pytest==3.3.2
pytest-html==1.16.1
pytest-metadata==1.5.1